+++ packages/injective-cosmwasm/src/exchange/mod.rs
@@ -2,6 +2,7 @@ pub mod derivative;
 pub mod derivative_market;
 pub mod order;
 pub mod privileged_action;
+pub mod response;
 pub mod spot;
 pub mod spot_market;
 pub mod subaccount;
