+++ packages/injective-cosmwasm/src/exchange/types.rs
@@ -7,13 +7,35 @@ use std::fmt;
 
 use crate::InjectiveQuerier;
 
+pub const UNSORTED_CANCELLATION_STRATEGY: i32 = 0;
+
+pub const FROM_WORST_TO_BEST_CANCELLATION_STRATEGY: i32 = 1;
+
 /// Deposit is data format for the subaccount deposit
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct Deposit {
-    #[serde(default)]
     pub available_balance: FPDecimal,
-    #[serde(default)]
     pub total_balance: FPDecimal,
+}
+
+/// Response to query for aggregate volume for a given market
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+pub struct DenomDecimals {
+    pub denom: String,
+    pub decimals: u64,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+pub struct PriceLevel {
+    pub p: FPDecimal,
+    pub q: FPDecimal,
+}
+
+impl PriceLevel {
+    // helper method for tests
+    pub fn new(p: FPDecimal, q: FPDecimal) -> PriceLevel {
+        PriceLevel { p, q }
+    }
 }
 
 /// Volume values divided by type (maker or taker volume)
