+++ packages/injective-cosmwasm/src/query.rs
@@ -1,4 +1,4 @@
-use cosmwasm_std::{Coin, CustomQuery, Uint128};
+use cosmwasm_std::CustomQuery;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
@@ -17,7 +17,6 @@ use crate::oracle::{
     volatility::{MetadataStatistics, TradeHistoryOptions, TradeRecord},
 };
 use crate::route::InjectiveRoute;
-use crate::wasmx::types::FundingMode;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -161,166 +160,3 @@ pub enum InjectiveQuery {
 }
 
 impl CustomQuery for InjectiveQueryWrapper {}
-
-pub const UNSORTED_CANCELLATION_STRATEGY: i32 = 0;
-
-pub const FROM_WORST_TO_BEST_CANCELLATION_STRATEGY: i32 = 1;
-
-/// SubaccountDepositResponse is data format returned from ExchangeQuery::SubaccountDeposit query
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct SubaccountDepositResponse {
-    pub deposits: Deposit,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct SubaccountEffectivePositionInMarketResponse {
-    pub state: Option<EffectivePosition>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct SubaccountPositionInMarketResponse {
-    pub state: Option<Position>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct PerpetualMarketInfoResponse {
-    pub info: Option<PerpetualMarketInfo>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct PerpetualMarketFundingResponse {
-    pub state: Option<PerpetualMarketFunding>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct TraderDerivativeOrdersResponse {
-    pub orders: Option<Vec<TrimmedDerivativeLimitOrder>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct TraderSpotOrdersResponse {
-    pub orders: Option<Vec<TrimmedSpotLimitOrder>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct DerivativeMarketResponse {
-    pub market: FullDerivativeMarket,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct SpotMarketResponse {
-    pub market: Option<SpotMarket>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct MarketMidPriceAndTOBResponse {
-    pub mid_price: Option<FPDecimal>,
-    pub best_buy_price: Option<FPDecimal>,
-    pub best_sell_price: Option<FPDecimal>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct PriceLevel {
-    pub p: FPDecimal,
-    pub q: FPDecimal,
-}
-
-impl PriceLevel {
-    // helper method for tests
-    pub fn new(p: FPDecimal, q: FPDecimal) -> PriceLevel {
-        PriceLevel { p, q }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct MarketVolatilityResponse {
-    pub volatility: Option<FPDecimal>,
-    pub history_metadata: Option<MetadataStatistics>,
-    pub raw_history: Option<Vec<TradeRecord>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct OracleVolatilityResponse {
-    pub volatility: Option<FPDecimal>,
-    pub history_metadata: Option<MetadataStatistics>,
-    pub raw_history: Option<Vec<TradeRecord>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryOrderbookResponse {
-    #[serde(default)]
-    pub buys_price_level: Vec<PriceLevel>,
-    #[serde(default)]
-    pub sells_price_level: Vec<PriceLevel>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct TokenFactoryDenomSupplyResponse {
-    pub total_supply: Uint128,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct TokenFactoryCreateDenomFeeResponse {
-    pub fee: Vec<Coin>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct RegisteredContract {
-    // limit of gas per BB execution
-    pub gas_limit: u64,
-    // gas price that contract is willing to pay for execution in BeginBlocker
-    pub gas_price: u64,
-    // is contract currently active
-    pub is_executable: bool,
-    // code_id that is allowed to be executed (to prevent malicious updates) - if nil/0 any code_id can be executed
-    pub code_id: Option<u64>,
-    // optional - admin addr that is allowed to update contract data
-    pub admin_address: Option<String>,
-    // optional -  address of the contract granting fee
-    // must be set if fund_mode is GrantOnly
-    pub granter_address: Option<String>,
-    /// funding mode
-    pub fund_mode: FundingMode,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryContractRegistrationInfoResponse {
-    pub contract: Option<RegisteredContract>,
-}
-
-/// Response to query for aggregate volumes of a given account/subaccount - divided by markets
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryAggregateVolumeResponse {
-    pub aggregate_volumes: Vec<MarketVolume>,
-}
-
-/// Response to query for aggregate volume for a given market
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryAggregateMarketVolumeResponse {
-    pub volume: VolumeByType,
-}
-
-/// Response to query for aggregate volume for a given market
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct DenomDecimals {
-    pub denom: String,
-    pub decimals: u64,
-}
-
-/// Response to query for decimals for a given denom
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryDenomDecimalResponse {
-    pub decimals: u64,
-}
-
-/// Response to query for decimals for multiple denoms
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryDenomDecimalsResponse {
-    pub denom_decimals: Vec<DenomDecimals>,
-}
-
-/// Response to query for fee multiplier for atomic order
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-pub struct QueryMarketAtomicExecutionFeeMultiplierResponse {
-    pub multiplier: FPDecimal,
-}
