+++ packages/injective-cosmwasm/src/tokenfactory/mod.rs
@@ -1,2 +1,1 @@
 pub mod response;
-pub mod types;
